Rename the `/scat-cs/bag` page to `/scat-cs/inspect`
