+++ app/src/app/scat-cs/inspect/ButtonClipboard.tsx
