+++ app/src/app/scat-cs/inspect/Chart.tsx
