+++ app/src/app/scat-cs/inspect/HowToCite.tsx
