+++ app/src/app/scat-cs/inspect/LinkToggle.tsx
