+++ app/src/app/scat-cs/inspect/Plot.tsx
