+++ app/src/app/scat-cs/inspect/PlotGrid.tsx
