+++ app/src/app/scat-cs/inspect/ReferenceList.tsx
