+++ app/src/app/scat-cs/inspect/Table.tsx
