+++ app/src/app/scat-cs/inspect/TableSelectable.tsx
