+++ app/src/app/scat-cs/inspect/TermsOfUseCheck.tsx
