+++ app/src/app/scat-cs/inspect/colors.ts
