+++ app/src/app/scat-cs/inspect/page.tsx
