+++ app/src/app/scat-cs/inspect/types.ts
