+++ app/src/pages/api/references/[format]/for-selection/[...ids].ts
@@ -42,7 +42,7 @@ const handler = createRouter<AuthRequest, NextApiResponse>()
         id: `LXCatData`,
         type: "dataset",
         title: "Data downloaded from the LXCat platform.",
-        URL: `${process.env.NEXT_PUBLIC_URL}/scat-cs/bag?ids=${
+        URL: `${process.env.NEXT_PUBLIC_URL}/scat-cs/inspect?ids=${
           unique_ids.join(",")
         }`,
       };
