+++ app/src/pages/api/scat-cs/bag/index.ts
@@ -25,12 +25,12 @@ const handler = createRouter<AuthRequest, NextApiResponse>()
     }
     const ids = idsSchema.parse(rawIds);
     const data = await byIds(ids);
-    data.url = `${process.env.NEXT_PUBLIC_URL}/scat-cs/bag?ids=${
+    data.url = `${process.env.NEXT_PUBLIC_URL}/scat-cs/inspect?ids=${
       ids.join(
         ",",
       )
     }`;
-    data.terms_of_use = `${process.env.NEXT_PUBLIC_URL}/scat-cs/bag?ids=${
+    data.terms_of_use = `${process.env.NEXT_PUBLIC_URL}/scat-cs/inspect?ids=${
       ids.join(",")
     }#terms_of_use`;
     res.json(data);
