+++ app/src/pages/api/scat-cs/bag/legacy.ts
@@ -33,9 +33,10 @@ const handler = createRouter<AuthRequest, NextApiResponse>()
       ([key, reference]) => [key, reference2bibliography(reference)],
     );
 
-    data.url = `${process.env.NEXT_PUBLIC_URL}/scat-cs/bag?ids=${idsString}`;
+    data.url =
+      `${process.env.NEXT_PUBLIC_URL}/scat-cs/inspect?ids=${idsString}`;
     data.terms_of_use =
-      `${process.env.NEXT_PUBLIC_URL}/scat-cs/bag?ids=${idsString}#terms_of_use`;
+      `${process.env.NEXT_PUBLIC_URL}/scat-cs/inspect?ids=${idsString}#terms_of_use`;
 
     const { convertMixture } = await import("@lxcat/converter");
     res.setHeader("Content-Type", "text/plain");
